Kafka setup with subscription

+++ backend/subscription/src/main/java/com/devision/subscription/dto/SearchProfileRequest.java
@@ -1,5 +1,6 @@
 package com.devision.subscription.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -9,7 +10,7 @@ public class SearchProfileRequest {
     public List<String> technicalTags;
     public List<String> employmentStatuses; // strings of enum names
     public String country;
-    public Integer minSalary;
-    public Integer maxSalary;
+    public BigDecimal minSalary;
+    public BigDecimal maxSalary;
     public String jobTitles; // semicolon-separated
 }
